advaced

+++ src/Controller/IncidenciaController.php
@@ -4,7 +4,9 @@ namespace App\Controller;
 
 use App\Entity\Incidencia;
 use App\Entity\Usuario;
+use App\Form\IncidenciaClientType;
 use App\Form\IncidenciaType;
+use App\Repository\ClienteRepository;
 use App\Repository\IncidenciaRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -43,11 +45,35 @@ class IncidenciaController extends AbstractController
         ]);
     }
 
+    #[Route('/registro/cliente/{id_cliente}', name: 'app_incidencia_new_with_client', methods: ['GET', 'POST'])]
+    public function newWithClient(Request $request, IncidenciaRepository $incidenciaRepository, ClienteRepository $clienteRepository ,string $id_cliente): Response
+    {
+        $incidencium = new Incidencia();
+        $cliente = $clienteRepository->find($id_cliente);
+        $form = $this->createForm(IncidenciaClientType::class, $incidencium, []);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $incidencium->setFecha(new \DateTime());
+            $incidencium->setUsuario($this->getCurrentUser());
+            $incidencium->setCliente($cliente);
+            $incidenciaRepository->add($incidencium);
+            return $this->redirectToRoute('cliente_detail', [ 'id' => $id_cliente ], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('incidencia/new.html.twig', [
+            'name_user'  => $this->getNameUser(),
+            'form' => $form,
+            'client_id' => $id_cliente
+        ]);
+    }
+
     #[Route('/{id}', name: 'app_incidencia_show', methods: ['GET'])]
     public function show(Incidencia $incidencium): Response
     {
         return $this->render('incidencia/show.html.twig', [
             'incidencium' => $incidencium,
+            'name_user'  => $this->getNameUser(),
         ]);
     }
 
@@ -72,8 +98,13 @@ class IncidenciaController extends AbstractController
     #[Route('/{id}/{id_cliente}', name: 'app_incidencia_delete', methods: ['POST'])]
     public function delete(Request $request, Incidencia $incidencium, IncidenciaRepository $incidenciaRepository, string $id_cliente): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         if ($this->isCsrfTokenValid('delete'.$incidencium->getId(), $request->request->get('_token'))) {
             $incidenciaRepository->remove($incidencium);
+            $this->addFlash(
+                'notice',
+                'Incidencia eliminada!'
+            );
         }
 
         return $this->redirectToRoute('cliente_detail', ['id' => $id_cliente], Response::HTTP_SEE_OTHER);
